+++ src/lua/CommunityBalanceMod/Minimap/CommandStation_Server.lua
@@ -1,10 +1,10 @@
 
-
-function CommandStation:OnEntityChange(oldId, newId)
+local oldCommandStructureOnEntityChange= CommandStructure.OnEntityChange
+function CommandStation:OnEntityChange(oldEntityId, _)
+    oldCommandStructureOnEntityChange(self, oldEntityId, _)
 
     if HasMixin(self, "MapBlip") then 
          self:MarkBlipDirty()
     end
-    --CommandStructure.OnEntityChange(self, oldId, newId)
 
 end