+++ src/lua/CommunityBalanceMod/Scripts/GUIBetaBalanceChangelogData.lua
@@ -79,9 +79,13 @@ gChangelogData =
 
 
   - Marines start with 2 IPs at 7+ players instead of 9+ players
-  - dropping mines cost 5 tres (from 7 tres)
-  - dropping welders cost 2 tres (from 3 tres)
-
+  - Dropping mines cost 5 tres (from 7 tres)
+  - Dropping welders cost 2 tres (from 3 tres)
+  - Autopickup for welders reduced from 5 to 1 second
+  - ARCs dont deal damage to other ARCS anymore
+  - Grenade launchers dont inflict selfdamage anymore
+  - Bile damage accelerates weapon expiration
+    - 1 Bile ~ 5 seconds
 
 ## QoL 
 ### Minimap rework
@@ -92,9 +96,12 @@ gChangelogData =
   - Added Icon for Drifter Eggs
   - Added Icon for deployed ARCS
   - Added Icon for Advanced Armory
+  - Marine structures under construction or powerless are slightly brighter
+
+  - Alien Commander is able to see parasited mines
+
 
 ## End of Playtest Changes
-
 
 
   #Changes between Revision 1.5.2 and Vanilla: